pyproject.toml
[project]
name = "rag-private-document-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.6.2",
    "chromadb>=1.4.1",
    "flask>=3.1.2",
    "flask-talisman>=1.1.0",
    "gunicorn>=24.1.1",
    "langchain==0.2.16",
    "langchain-community==0.2.16",
    "langchain-huggingface==0.0.3",
    "langchain-openai==0.1.23",
    "langsmith>=0.1.147",
    "prometheus-flask-exporter>=0.23.2",
    "pydantic-settings>=2.12.0",
    "pypdf>=6.6.2",
    "python-dotenv>=1.2.1",
    "redis>=7.1.1",
    "sentence-transformers>=5.2.2",
    "structlog>=25.5.0",
]

[project.optional-dependencies]
dev = [
    "pylint>=3.0.0",
    "pytest>=8.0.0",
    "black>=24.0.0",
    "mypy>=1.8.0",
    "bandit>=1.7.0",
    "pytest-cov>=4.1.0",
]

[tool.pytest.ini_options]
pythonpath = "."
addopts = "--cov=src --cov-report=term-missing"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.*"
ignore_errors = true

[tool.bandit]
exclude_dirs = ["tests", ".venv"]
skips = ["B101"]

[tool.coverage.run]
source = ["src"]
omit = ["src/original_main.py"]

[tool.black]
line-length = 88
